move component into contents folder
