+++ src/components/ui/content/textCenter.tsx
